chore: Setup new contact fields
